% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anstrain}
\alias{anstrain}
\title{Perceptual learning data}
\format{
An object of class \code{data.frame} with 1000 rows and 6 columns.
}
\usage{
anstrain
}
\description{
Data from 4 participants ("subID") completing 250 trials each of perceptual learning.
"resp" is participant response to the "ratio" offset, with correctness recorded
as "acc". "sizeRat" is a covariate.
}
\keyword{datasets}
