% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anstrain_s1}
\alias{anstrain_s1}
\title{Perceptual learning data from one participant}
\format{
An object of class \code{data.frame} with 250 rows and 6 columns.
}
\usage{
anstrain_s1
}
\description{
Subset of data from Cochrane, Cui, Hubbard & Green (2019; DOI: 10.3758/s13414-018-01636).
Data from 1 participant completing 250 trials of perceptual learning in an
Approximate Number System ratio discrimination task. "resp" is participant response to the
"ratio" offset of two interspersed colors of dots, with correctness recorded as "acc".
"sizeRat" is the ratio of sizes of the different colors of dots.
}
\keyword{datasets}
