% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.TEfit.R
\name{coef.TEfit}
\alias{coef.TEfit}
\title{Extract coefficients from a TEfit model}
\usage{
\method{coef}{TEfit}(TEs3)
}
\arguments{
\item{TEs3}{A TEfit model, possibly with resampled parameters}
}
\description{
Extract coefficients from a TEfit model
}
