% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.TEfitAll.R
\name{fitted.TEfitAll}
\alias{fitted.TEfitAll}
\title{Get fitted values and summary statistics from a set of TEfit models}
\usage{
\method{fitted}{TEfitAll}(TEs3s)
}
\arguments{
\item{TEs3s}{A set of models fit by TEfitAll()}
}
\description{
Get fitted values and summary statistics from a set of TEfit models
}
