% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_logit.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse logistic transformation}
\usage{
inv_logit(q)
}
\arguments{
\item{q}{}
}
\description{
This is simply `stats::plogis`, renamed for compatibility between R and 
Stan. `brms` previously included this, but it was removed from that package.
}
\examples{

all.equal(inv_logit(2),plogis(2))

}
