% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TEfit.R
\name{plot.TEfit}
\alias{plot.TEfit}
\title{Plot a TEfit}
\usage{
\method{plot}{TEfit}(TEs3, plot_title = "", xlabel = "", ylabel = "",
  sub_title = "", ymin = NA, ymax = NA)
}
\arguments{
\item{TEs3}{TEfit model output}

\item{plot_title}{optional plot title}

\item{xlabel}{optional plot x axis label}

\item{ylabel}{optional plot y axis label}

\item{sub_title}{optional plot caption}

\item{ymin}{optional lower boundary for Y axis}

\item{ymax}{optional upper boundary for X axis}
}
\description{
Plot predicted values of a TEfit model. Predictions are thresholds, if relevant;
otherwise they are overall model predictions
}
