% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TEfitAll.R
\name{plot.TEfitAll}
\alias{plot.TEfitAll}
\title{Plot a collection of models fit by TEfitAll()}
\usage{
\method{plot}{TEfitAll}(TEs3s, ymin = NA, ymax = NA)
}
\arguments{
\item{TEs3s}{A TEfitAll model object (collection of fit models)}

\item{ymin}{minimum y value}

\item{ymax}{maximum y value}
}
\description{
Plot a collection of models fit by TEfitAll()
}
\examples{
\dontrun{
m <- TEfitAll(anstrain[,c('acc','trialNum')],groupingVar = anstrain$subID)
plot(m)
}

}
