% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.TEfit.R
\name{residuals.TEfit}
\alias{residuals.TEfit}
\title{Extract residuals from a TEfit model}
\usage{
\method{residuals}{TEfit}(modelIn, toPlot = F)
}
\arguments{
\item{modelIn}{A TEfit model}
}
\description{
Deprecated
}
