% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.TEfit.R
\name{simulate.TEfit}
\alias{simulate.TEfit}
\title{Simulate from a TEfit, assuming multivariate gaussian parameters}
\usage{
\method{simulate}{TEfit}(model_in, nsim = 100, newdata = data.frame(),
  toPlot = F)
}
\arguments{
\item{model_in}{A TEfit model}

\item{nsim}{number of simulations to generate}

\item{newdata}{If desired, new data from which to generate values}

\item{toPlot}{Logical, for plotting of values.}
}
\description{
Requires the $psych$ and $MASS$ packages.
}
