% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TEfit.R
\name{summary.TEfit}
\alias{summary.TEfit}
\title{Summarize a time-evolving fit}
\usage{
\method{summary}{TEfit}(TEs3, printOutput = T)
}
\arguments{
\item{TEs3}{A fit TEfit model}

\item{printOutput}{Print output to console (if T) or return a list of summary items (if F)}
}
\description{
Prints or returns a summary of a TEfit object. This includes parameter values,
convergence, the full formula, goodness-of-fit metrics, measures of change in
conditional independence, and (if applicable) distributional information from resampling.
}
\details{
Pseudo-SE is an approximation to the standard error of the parameter using bootstrapped estimates. It is calculated by
first calculating the .025 and .975 quantile() of the resampled parameters. Then, the absolute
difference is calculated bewteen each of these CI values and the overall fit value. The absolute
differences are divided by qnorm(.975) and averaged in order to get the pseudo-SE, the
pseudo standard deviation of the parameter expected value.
}
