% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.TEfitAll.R
\name{summary.TEfitAll}
\alias{summary.TEfitAll}
\title{Summarize a collection of time-evolving fits}
\usage{
\method{summary}{TEfitAll}(TEs3s, printOutput = T, printAll = F)
}
\arguments{
\item{TEs3s}{A set of fit TEfit models}

\item{printOutput}{Print output to console (if T) or return a list of summary items (if F)}

\item{printAll}{Print grouping-level fits?}
}
\description{
Summarize a collection of time-evolving fits
}
