% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEbrm.R
\name{TEbrm}
\alias{TEbrm}
\title{Fit a time-evolving model with Stan using brms}
\usage{
TEbrm(
  formIn,
  dataIn,
  iter = 1000,
  chains = 3,
  priorIn = c(),
  algorithm = "sampling",
  ...,
  tef_control_list = TEfits::tef_control()
)
}
\arguments{
\item{formIn}{Formula to fit. See examples.}

\item{dataIn}{Data frame, from which to fit the model.}

\item{iter}{Number of iterations to run the model.}

\item{chains}{Number of chains to run the model.}

\item{priorIn}{Optional argument to pass priors to the \code{brms} model, alongside the TEfit-default rate prior.}

\item{algorithm}{The algorithm to use when fitting the \code{\link[brms]{brm}} model}

\item{...}{Further arguments passed to the brms model}

\item{tef_control_list}{A list of control parameters passed in by \code{tef_control()}}
}
\description{
Formats and runs a \code{\link[brms]{brm}s} model for a time-evolving nonlinear function. Function is \strong{under development}
and is likely to be buggy, and to change frequently.
}
\examples{
\dontrun{
## Default model formula is exponential change, with no covariates or random effects
m <- TEbrm(
acc ~ trialNum
,priorIn = prior(normal(.5,.5),nlpar='pAsym') + prior(normal(.5,.5),nlpar='pStart')
,dataIn = anstrain_s1
)

summary(m)
conditional_effects(m)

## using the tef_change_expo3 function to construct the model formula, with random effects
m <- TEbrm(
acc ~ tef_change_expo3('trialNum',parForm = ~ (1|subID))
,dataIn = anstrain
,priorIn = prior(normal(.5,.5),nlpar='pAsym') + prior(normal(.5,.5),nlpar='pStart')
)

}
}
\seealso{
For additional flexibility, and full explanations of model options, see \code{\link[brms]{brms-package}}.
}
