% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_acc2dprime.R
\name{tef_acc2dprime}
\alias{tef_acc2dprime}
\title{Convert an ordered vector of accuracies to d-prime}
\usage{
tef_acc2dprime(accuracy, stim_present, by_index = T, trial_hwhm = 2,
  max_dprime = 5)
}
\arguments{
\item{accuracy}{At each index, what was the accuracy: [bounded at 0 and 1]}

\item{stim_present}{At each index, was the stimulus present or absent: [binary; 0 and 1, or logical]}

\item{by_index}{Should the d-prime be calculated for each index?}

\item{trial_hwhm}{The Gaussian smoother has a half-width-half-max; values are given half weight at this index distance from the center index (as the smoother iterates through each index in turn as the center).}

\item{max_dprime}{d-prime becomes infinite as accuracy approaches 0 or 1. This value limits the absolute value of d-prime.}
}
\description{
Given paired information regarding accuracy and stimulus presence,
run a Gaussian-weighted-mean smoother over the accuracy vector separately for
stimulus-present and stimulus-absent indices, then compute an index-wise d-prime.
Returning and entire-vector d-prime (i.e., stable across time) is also an option.
}
