% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_applyWDM.R
\name{tef_applyWDM}
\alias{tef_applyWDM}
\title{Calculate the Wiener negative log likelihood of a vector of data using a vector of drift rates.}
\usage{
tef_applyWDM(dat, DR, BS, NDT, Bias = 0.5)
}
\arguments{
\item{dat}{Vector of data (e.g., RT). Hitting upper bound must be positively signed; hitting lower bound must be negatively signed.}

\item{DR}{Vector of drift rates. Must be the same length as dat.}

\item{BS}{Boundary separation parameter. Positive scalar.}

\item{NDT}{Non-decision time parameter. Positive scalar.}

\item{Bias}{Bias parameter. Positive scalar.}
}
\description{
Within a TEfit call, errFun='wiener_df' is likely to be extremely slow, e.g., 70 seconds per run.
}
\examples{
dat <- c(-1,1,-.5,.5)
driftRate <- c(.2,.3,.4,.5)
tef_applyWDM(dat,DR=driftRate,BS=1,NDT=.3,Bias=.5)

}
