% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_bootFits.R
\name{tef_bootFits}
\alias{tef_bootFits}
\title{Fit a model to resampled data}
\usage{
tef_bootFits(modList)
}
\arguments{
\item{modList}{List of TEfit model details}
}
\description{
\code{\link{TEfit}} internal. Given the modList constructed in TEfit(), refit the model nBoots number of times
on bootPercent proportion of the data. If bootPercent is 1, sampling is done with replacement.
If bootPercent is between 0 and 1, sampling is done without replacement, i.e., a random
subset of the data is selected. This latter case allows train/test assessment. That is,
the model is fit to some subset of the data, after which the predicted values for the remaining data
are calculated. The error of these out-of-sample predictions are recorded alongside the
in-sample goodness-of-fit.
}
\details{
Number of runs per resample [bootTries] defaults to 10, the minimum.
}
