% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_bootList.R
\name{tef_bootList}
\alias{tef_bootList}
\title{Make a list of parameters for resampling a TEfit model}
\usage{
tef_bootList(resamples = 0, bootPercent = 1, bootTries = 20)
}
\arguments{
\item{resamples}{Number of resamples. More (e.g., >2000) is better, but requires more time and memory.}

\item{bootPercent}{Proportion of data to resample. Must be 0 < bootPercent <= 1. If 1, data is resampled repeatedly with replacement. If less than 1, that proportion of data is repeatedly randomly selected, the model is fit on the selected data, and fit indices (e.g., error) are calculated for the model predictions on the left-out data (i.e., cross-validation).}

\item{bootTries}{Number of optimization runs for each resample}
}
\description{
Used to assist the user in the specification of a set of parameters for resampling analyses.
}
