% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_changeFun.R
\name{tef_changeFun}
\alias{tef_changeFun}
\title{Construct the right hand side of the equation for a TEfit}
\usage{
tef_changeFun(modList, covarVects)
}
\arguments{
\item{modList}{List of TEfit model details}

\item{covarVects}{Vectors of covariate(s) for each model parameter}
}
\description{
Construct the right hand side of the equation for a TEfit
}
