% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_checkPars.R
\name{tef_checkPars}
\alias{tef_checkPars}
\title{Check for bound hitting and other undesireable outcomes within an optim() call}
\usage{
tef_checkPars(err, guesses, curDat, pNames, evalFun, errFun, respVar,
  linkFunX = NA, y_lim, rate_lim, shape_lim, penalizeRate, paramTerms,
  guessGroups = guessGroups)
}
\arguments{
\item{err}{?}

\item{guesses}{?}

\item{curDat}{?}

\item{pNames}{?}

\item{evalFun}{?}

\item{errFun}{?}

\item{respVar}{?}

\item{y_lim}{?}

\item{rate_lim}{?}

\item{shape_lim}{?}

\item{penalizeRate}{?}

\item{paramTerms}{?}

\item{guessGroups}{?}
}
\description{
Check for bound hitting and other undesireable outcomes within an optim() call
}
