% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_control.R
\name{tef_control}
\alias{tef_control}
\title{Get control parameters for a TEfit model}
\usage{
tef_control(quietErrs = F, suppressWarnings = F, nTries = 200,
  y_lim = c(-1e+07, 1e+07), rate_lim = c(0, 0), shape_lim = c(0, 0),
  expBase = 2, rateBase = 2, pFix = c(), penalizeMean = T,
  penalizeRate = F, convergeTol = 0.05, stepwise_asym = F,
  explicit = "")
}
\arguments{
\item{quietErrs}{..}

\item{suppressWarnings}{..}

\item{y_lim}{..}

\item{rate_lim}{..}

\item{expBase}{..}

\item{rateBase}{..}

\item{pFix}{..}

\item{penalizeMean}{..}

\item{penalizeRate}{..}

\item{convergeTol}{..}

\item{stepwise_asym}{..}

\item{explicit}{..}

\item{modList}{..}
}
\description{
Get control parameters for a TEfit model
}
