% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_err.R
\name{tef_err}
\alias{tef_err}
\title{Calculate error for TEFit::tef_fitErr}
\usage{
tef_err(y, yHat, errFun, curDat = NA)
}
\arguments{
\item{y}{Actual response variable values}

\item{yHat}{Predicted response variabel values}

\item{errFun}{Error function (e.g., OLS, logcosh)}

\item{curDat}{Data input. Only really relevant when calculating a full likelihood function and needing info beyond yHat}
}
\description{
\code{\link{TEfit}} internal
}
