% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_fitErr.R
\name{tef_fitErr}
\alias{tef_fitErr}
\title{Calculate the error of data, given the model and certain parameters}
\usage{
tef_fitErr(guesses, varIn, pNames, evalFun, errFun, respVar,
  linkFunX = NA, y_lim, rate_lim, shape_lim, penalizeRate, parLims,
  thresh_fun, paramTerms)
}
\arguments{
\item{guesses}{proposed parameter values}

\item{varIn}{Data to be fit}

\item{pNames}{Parameter names}

\item{evalFun}{Model function}

\item{errFun}{Error function}

\item{respVar}{The name of the response variable}

\item{y_lim}{Bounds for the predicted values}

\item{rate_lim}{Bounds for the rate parameter}

\item{shape_lim}{Bounds for the shape parameter, if needed.}

\item{penalizeRate}{logical, whether or not to penalize rate parameter}

\item{parLims}{parameter limits}

\item{thresh_fun}{if relevant, the threshold function}

\item{paramTerms}{the terms for parameters}
}
\description{
This function should be improved. It is very very slow currently,
and its use of grep() and if() statements should be minimized or removed.
A bit of creativity will be needed for the
reaching-boundaries and linkfun-specific checks
}
