% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getLinkedFun.R
\name{tef_getLinkedFun}
\alias{tef_getLinkedFun}
\title{Gets link for for TEFit}
\usage{
tef_getLinkedFun(modList)
}
\arguments{
\item{modList}{List of TEfit model details}
}
\description{
\code{\link{TEfit}} internal. This function also calculates default rate and [weibull] shape bounds.
}
