% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getRanefForm.R
\name{tef_getRanefForm}
\alias{tef_getRanefForm}
\title{Extract group-level regressions from a multilevel model formula}
\usage{
tef_getRanefForm(formIn, groupingVar)
}
\arguments{
\item{formIn}{multilevel model formula}

\item{groupingVar}{grouping variable that will be the left-hand-side of the returned formula}
}
\description{
Returns the groupingVar-level regression formulas. For example, the
formula \code{yVar ~ aVar \* bVar \* cVar + (aVar \* cVar | groupingVar)}
will return the group-level formula \code{groupingVar ~ aVar \* cVar}. Note that
All random effects of \code{groupingVar} must be within a single set of (parentheses).
}
