% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsExpo.R
\name{tef_getTermsExpo}
\alias{tef_getTermsExpo}
\title{{{CREATE EXPLANATION}}}
\usage{
tef_getTermsExpo(modList, pPrefix)
}
\arguments{
\item{modList}{{{CREATE EXPLANATION}}}

\item{pPrefix}{{{CREATE EXPLANATION}}}
}
\description{
Note that, despite the name, this is the getterms function for  *both* exponential and power [3-parameter] change functions
}
