% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsExpoBlock.R
\name{tef_getTermsExpoBlock}
\alias{tef_getTermsExpoBlock}
\title{{{CREATE EXPLANATION}}}
\usage{
tef_getTermsExpoBlock(modList, pPrefix)
}
\arguments{
\item{modList}{List of TEfit model details}

\item{pPrefix}{{{CREATE EXPLANATION}}}
}
\description{
pBlScale indicates the warping of distance-to-asymptote based on block trial number.
}
