% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsExpoDouble.R
\name{tef_getTermsExpoDouble}
\alias{tef_getTermsExpoDouble}
\title{{{CREATE EXPLANATION}}}
\usage{
tef_getTermsExpoDouble(modList, pPrefix)
}
\arguments{
\item{modList}{{{CREATE EXPLANATION}}}

\item{pPrefix}{{{CREATE EXPLANATION}}}
}
\description{
{{CREATE EXPLANATION}}
}
