% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsExpoFatigue.R
\name{tef_getTermsExpoFatigue}
\alias{tef_getTermsExpoFatigue}
\title{{{CREATE EXPLANATION}}}
\usage{
tef_getTermsExpoFatigue(modList, pPrefix)
}
\arguments{
\item{modList}{{{CREATE EXPLANATION}}}

\item{pPrefix}{{{CREATE EXPLANATION}}}
}
\description{
Exponential change followed by logistic decay.
}
