% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsPower4.R
\name{tef_getTermsPower4}
\alias{tef_getTermsPower4}
\title{{{CREATE EXPLANATION}}}
\usage{
tef_getTermsPower4(modList, pPrefix)
}
\arguments{
\item{modList}{{{CREATE EXPLANATION}}}

\item{pPrefix}{{{CREATE EXPLANATION}}}
}
\description{
{{CREATE EXPLANATION}}
}
