% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsWeibull.R
\name{tef_getTermsWeibull}
\alias{tef_getTermsWeibull}
\title{Define the terms for a Weibull change function}
\usage{
tef_getTermsWeibull(modList, pPrefix)
}
\arguments{
\item{modList}{TEfit modList}

\item{pPrefix}{prefix for parameters}
}
\description{
Internal to TEfits.
}
