% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_link_logistic.R
\name{tef_link_logistic}
\alias{tef_link_logistic}
\title{Construct a logistic link function parameterized with threshold and bias}
\usage{
tef_link_logistic(
  changeForm,
  linkX,
  changePar = c("threshold", "bias"),
  threshVal = 0.75,
  lapseRate = 0.005,
  boundScale = 2,
  constantPar_prior = "normal(0,3)"
)
}
\arguments{
\item{changeForm}{The formula describing the change in either}

\item{linkX}{Character. The name of the "x" variable in the logistic link function (e.g., stimulus strength in a psychometric function)}

\item{changePar}{Character. Which variable, "threshold" or "bias", changes over time. The other one is stable over time; \emph{the stable component inherits its formula from the asymptote component}.}

\item{threshVal}{The threshold at which to evaluate the logistic function (i.e., the y-value for which threshold describes the x-value).}

\item{lapseRate}{The offset, from 0 or 1, of the logistic function at arbitrarily large (positive or negative) values of \code{linkX}. A small lapse rate improves model fit (see Wichmann and Hill, 2001, P&P).}

\item{boundScale}{Currently not implemented. Upper threshold of threshold estimates, as a multiple of the maximum absolute \code{linkX}.}

\item{constantPar_prior}{The prior to put on the constant component of the logistic function (i.e., either the bias or the log threshold). Only relevant if passing the model to a \code{\link[brms]{brm}} model (e.g., with \code{\link{TEbrm}}).}
}
\description{
Function is \strong{under development}
and is likely to be buggy, and to change frequently.
}
\details{
If bias is changing, threshold inherits its formula from asymptotic bias.
If threshold is changing, bias inherits its formula from asymptotic threhold.
}
\examples{

equation_to_fit <- tef_link_logistic( tef_change_expo3('trialNum', parForm = ~ (1|subID)) , linkX = 'ratio' )

}
