% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_link_weibull.R
\name{tef_link_weibull}
\alias{tef_link_weibull}
\title{Construct a weibull link function parameterized with threshold and bias}
\usage{
tef_link_weibull(
  changeForm,
  linkX,
  threshVal = 0.75,
  rhAsymptote = 1,
  yIntercept = 0.5,
  lapseRate = 0.005,
  boundScale = 2,
  constantPar_prior = "normal(0,3)"
)
}
\arguments{
\item{changeForm}{The formula describing the change in either}

\item{linkX}{Character. The name of the "x" variable in the weibull link function (e.g., stimulus strength in a psychometric function)}

\item{threshVal}{The threshold at which to evaluate the weibull function (i.e., the y-value for which threshold describes the x-value).}

\item{rhAsymptote}{The asymptotic value of the weibull function with large \code{linkX} values (e.g., accuracy at infinitely large stimulus strength).}

\item{yIntercept}{The origin value of the weibull function (with a \code{linkX} value of zero, e.g., accuracy at a stimulus strength of zero; in behavioral data is likely to be "guessing rate").}

\item{lapseRate}{The offset, from rhAsymptote, of the weibull function at arbitrarily large values of \code{linkX}. A small lapse rate improves model fit (see Wichmann and Hill, 2001, P&P).}

\item{boundScale}{Currently not implemented. Upper threshold of threshold estimates, as a multiple of the maximum absolute \code{linkX}.}

\item{constantPar_prior}{The prior to put on the constant component of the weibull function (i.e., shape parameter). Only relevant if passing the model to a \code{\link[brms]{brm}} model (e.g., with \code{\link{TEbrm}}).}
}
\description{
Function is \strong{under development}
and is likely to be buggy, and to change frequently.
}
\details{
\strong{\code{shape}} is a parameter in the weibull function, and \strong{\code{must not}} be
used as a name of data variables (e.g., \code{linkX} or within \code{changeForm}).
}
\examples{

equation_to_fit <- tef_link_weibull( tef_change_expo3('trialNum', parForm = ~ (1|subID)) , linkX = 'absoluteRatio' )

}
