% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_mono_expo.R
\name{tef_mono_expo}
\alias{tef_mono_expo}
\title{Create an integer vector coding a saturating function of time}
\usage{
tef_mono_expo(N, style = "increasing")
}
\arguments{
\item{N}{Number of time values}

\item{style}{Should the integer "blocks" be saturating (increasing in length) ("increasing") or equal in size ("equal")}
}
\description{
Create an integer vector coding a saturating function of time
}
\examples{
\dontrun{
# # Make a temporary data set
d_tmp <- anstrain_s1

# # Get an integer-coded time variable
d_tmp$trialNum_mono <- tef_mono_expo(nrow(d_tmp))

# # Look at the new coding of time
plot(d_tmp$trialNum,d_tmp$trialNum_mono)

# # Fit a model predicting accuracy using a monotonic measure of time; see documentation for brms::mo
library(brms)
m <- brm(acc ~ mo(trialNum_mono), d_tmp)
}
}
