% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_parseParFormula.R
\name{tef_parseParFormula}
\alias{tef_parseParFormula}
\title{Parse a parameter-level formula for a TEfit model}
\usage{
tef_parseParFormula(parFormula, label = "pPar")
}
\arguments{
\item{parFormula}{Formula for fitting a parameter. Variables should be numeric; non-numeric variables may break. Can also take a string or a scalar, if the parameter is to be fixed to that value.}

\item{label}{Label for the parameters}
}
\description{
Returns the \code{parFormula}, with some additional information that assists TEfits in fitting.
}
\details{
If there are no random-effects terms, parses into a literal equation (e.g., turns 
\code{~ x1 + x2 * x3} into \code{pPar_Intercept + pPar_x1\*x1 + pPar_x2\*x2 + pPar_x3\*x3 + pPar_x2_x3\*x2\*x3}) 
which is returned as the "equation" attribute that has an accompanying "parameters" attribute. Additional attributes
include whether there are random-effects terms (attribute "MEM") and whether there are any parameters at all, or if the
"equation" is actually fixed to a variable or scalar (attributed "is_fixed").
}
\examples{
tef_parseParFormula(~ x1 + x2*x3 + (x2 || subID))
tef_parseParFormula(~ x1 + x2*x3, label = 'thresholdAsymptote')
tef_parseParFormula(2)

}
