% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_penalizedErr.R
\name{tef_penalizedErr}
\alias{tef_penalizedErr}
\title{Penalize error for being close to a boundary}
\usage{
tef_penalizedErr(boundedPar, errPar, loBound, upBound,
  dbeta_penalty = 1.001)
}
\arguments{
\item{boundedPar}{Value of the parameter that has bounds.}

\item{errPar}{Value of the error to be penalized.}

\item{loBound}{Lower parameter boundary.}

\item{upBound}{Upper parameter boundary.}

\item{dbeta_penalty}{The multiplicative error penalty at approximately 5.3 percent away from the boundary. Error increases with increasing proximity to a bound.}
}
\description{
Formally, `pErr=Err/dbeta(par,dbeta_penalty,dbeta_penalty)`, where
par is the rate parameter normalized between 0 [lower bound] and 1 [upper bound]
}
