% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_quantileBoot.R
\name{tef_quantileBoot}
\alias{tef_quantileBoot}
\title{Bootstrap quantiles}
\usage{
tef_quantileBoot(x, probs = c(0.025, 0.5, 0.975), nBoot = 1000, na.rm = T, ...)
}
\arguments{
\item{x}{Numeric vector over which to resample and compute quantiles}

\item{probs}{Numeric vector of probabilities with values in \emph{[0,1]}.}

\item{nBoot}{Number of resamples-with-replacement}

\item{na.rm}{Logical; if true, any NA and NaN's are removed from \code{x} before the quantiles are computed.}

\item{...}{Other arguments to pass to \code{quantile()}}
}
\description{
Resamples a numeric vector with replacement, finds the requested quantiles
for each resample, and returns the trimmed average of the resampled quantiles.
See \code{\link[stats]{quantile}}.
}
\details{
Returns the mean of the middle 5% of the resamples for each quantile (i.e., \code{mean(quantile_i,trim = .475)}). In
practice this means that the median will be returned if the number of resamples is small enough, and
that the default number of 1000 resamples will use the middle 50 resamples to compute the mean.
}
\examples{
exp_quantiles <- tef_quantileBoot(c(rexp(100),NA))
cat(exp_quantiles)

}
