% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_tryFits.R
\name{tef_tryFits}
\alias{tef_tryFits}
\title{Fit a TE model}
\usage{
tef_tryFits(modList, whichPnames = "pNames", whichFun = "evalFun")
}
\arguments{
\item{modList}{List of TEfit model details}

\item{whichPnames}{Used to distinguish between the TE and null models}

\item{whichFun}{Used to distinguish between the TE and null models.}
}
\description{
\code{\link{TEfit}} internal
}
\details{
This function takes a modList, formats inputs for tef_fitErr(),
and uses optim() to find many fits of the modList. Then all of the
fits are compared in order to determine which are best and whether the
10 best fits have similar parameters (SD<convergeTol).
}
