% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_vars2forms.R
\name{tef_vars2forms}
\alias{tef_vars2forms}
\title{Get the TEfit model formula}
\usage{
tef_vars2forms(modList)
}
\arguments{
\item{modList}{List of TEfit model details}
}
\description{
Get the TEfit model formula
}
