% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_windowWDM.R
\name{tef_windowWDM}
\alias{tef_windowWDM}
\title{Find a data frame of Wiener Diffusion Model parameters}
\usage{
tef_windowWDM(dat, windowSize = 20, fit_BS = FALSE)
}
\arguments{
\item{dat}{Response time vector. Positive values indicates upper-boundary while negative values indicate lower-boundary.}

\item{windowSize}{Total width of each window for estimating parameters.}

\item{fit_BS}{Logical. If \code{FALSE} only drift rate is estimated, if \code{TRUE} both drift rate and boundary separation are estimated.}
}
\description{
Estimates a set of Wiener Diffusion Model parameters for a vector of formatted
data. Estimation uses \code{\link[RWiener]{wdm}} and data must be formatted as a
1-dimensional vector according the to \code{RWiener} package (i.e., RT * 2 * (boundary=='upper' - 0.5) ).
Assumes that RTs are associated with equally-spaced and sorted timepoints (e.g., trials of a behavioral study).
First fits a \code{\link[RWiener]{wdm}} for the entire RT vector, and applies some of these parameters
(NDT, bias, and possibly BS) to the entire RT vector. Next takes a set of windows, equally spaced and centered at
\code{windowSize/2}, and estimates the DR (and possibly BS) for each window. Averages overlapping windows.
Returns a data frame of estimated parameters, including 95\% CI.
}
\examples{
\dontrun{
dat <- (rnorm(200,.2,.05) + rexp(200,1 / .2))*sample(c(-1,1),200,replace=T)
    m_drY_bsN <- tef_windowWDM(dat,fit_BS = F)
    m_drY_bsY <- tef_windowWDM(dat,fit_BS = T)
    psych::pairs.panels(cbind(trialNum = 1:200, RT = abs(dat),m_drY_bsY[,1:6]))
    }
}
