% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEfitAll.R
\name{TEfitAll}
\alias{TEfitAll}
\title{Fit several time-evolving regression models}
\usage{
TEfitAll(
  varIn,
  groupingVar,
  groupingVarName = "grouping_var",
  returnAll = T,
  progressDot = T,
  linkFun = list(link = "identity"),
  errFun = "ols",
  changeFun = "expo",
  bootPars = tef_bootList(),
  blockTimeVar = NULL,
  covarTerms = list(),
  control = tef_control()
)
}
\arguments{
\item{varIn}{Data frame or vector. First column [or vector] must be the time-dependent response variable (left hand side of regression). If available, second column must be the time variable. All other columns are covariates, possibly involved in a link function.}

\item{groupingVar}{Variable (e.g., participant ID) with which to separate TEfit models. Length must be nrows(varIn)}

\item{groupingVarName}{Name of grouping var}

\item{returnAll}{Logical. Return only a summary (when T), or that summary plus every model, in a list (when F)}

\item{progressDot}{If TRUE, prints a dot after each group fit}

\item{linkFun}{A list defining a link function (i.e., 'identity', 'd_prime', 'weibull', or 'logistic')}

\item{errFun}{A string defining an error function (e.g., 'ols', 'logcosh', 'bernoulli').}

\item{changeFun}{A string defining the functional form of change (e.g., 'expo', 'power', 'weibull')}

\item{bootPars}{A list defining the details for bootstrapped fits. Defaults to no bootstrapping. Necessary for estimates of uncertainty around fits and for covariance between parameters.}

\item{blockTimeVar}{A string identifying which covariate is the time points of sub-scales (e.g., "blocks" of times within the overall timescale of data collection)}

\item{covarTerms}{An optional list of logical vectors indicating whether parameters should vary by covariates. See examples.}

\item{control}{A list of model parameters. Use of tef_control() is highly recommended.}
}
\description{
A wrapper for fitting a \code{\link{TEfit}} model to the data
for every unique value of groupingVar. Defaults to
returning a list including two summaries and all models; returning
only a summary is also an option. Arguments (except groupingvar, a grouping vector)
are identical to, and are passed directly to, \code{\link{TEfit}}.
}
\examples{
\dontrun{
m <- TEfitAll(anstrain[,c('acc','trialNum')],groupingVar = anstrain$subID,groupingVarName = 'subID',bootPars = tef_bootList(resamples = 20))
summary(m)
}

}
\seealso{
\code{\link{TEfit}} for fitting a single model;
\code{\link{tef_fitAll2brms}} to re-fit the TEfitAll output using \code{\link[brms]{brms}}
}
