% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TElmem.R
\name{TElmem}
\alias{TElmem}
\title{Linear mixed-effects model with nonlinear time random effects}
\usage{
TElmem(
  formIn,
  dat,
  timeVar,
  groupingVar,
  nRuns = 1,
  startingOffset = T,
  silent = F
)
}
\arguments{
\item{formIn}{model formula, as in \code{lmer}}

\item{dat}{model data, as in \code{lmer}}

\item{timeVar}{String. Indicates which variable in \code{datIn} corresponds to time (i.e., should be transformed). Must be numeric and positive.}

\item{groupingVar}{String. Indicates which variable in \code{datIn} should have a time=related random effect.}

\item{nRuns}{Number of times to run optimization of the rate (i.e., fitting nonlinear transformations of \code{timeVar})}

\item{startingOffset}{By default (if T) time is coded to start at 1 and saturate to 0. If startingOffset is F, time starts at 0 and saturates to 1. May assist in interpreting interactions with other variables, etc.}

\item{silent}{Progress is printed by default. silent=T to suppress}
}
\value{
A list including:
\describe{
\item{\code{lmerMod}}{\code{\link[lme4]{lmer}} model fit with transformed time variable}
\item{\code{rates}}{Named vector of rates [\emph{50-percent-of-change time constants}]}
\item{\code{timeDat}}{Data frame with original and transformed time variable}
\item{\code{groupMods}}{List of fit \code{\link{TElm}} models, and the corresponding transformed time variable and named vector of rates}
}
}
\description{
Fits a \code{\link[lme4]{lmer}} linear mixed-effects model with the random effects of
\code{timeVar} for each level of \code{groupingVar}. Provides estimates of time-related change
(i.e., attempts to answer the question "how different was the start than the end?").
}
\details{
First uses \code{\link{TElm}} to find a rate parameter for each level of \code{groupingVar}, with
the formula extracted using \code{\link{tef_getRanefForm}}. These
rate parameters are used to transform the corresponding \code{timeVar} into a exponentially
saturating variable (see \code{\link{TElm}}). After finding an initial set of
rate parameters using \code{\link{TElm}},
\code{TElmem} attempts to optimize the vector of rate parameters in conjunction with the full
\code{lmer} model.

May be used, with \code{nRuns=0}, to simply
use rate estimates from independent \code{groupingVar}-level \code{\link{TElm}} models, extracting the corresponding
transformed time variables, and using them in a LMEM.
}
\note{
Random effects and rate estimates may be unstable, and optimization may take
a very long time to run. The primary purpose of this function is to allow for by-\code{groupingVar}
detrending of time-related changes in data (i.e., to estimate and test fixed effects at asymptotic time,
or to estimate and test the magnitude of time-related effects).
If reliable by-\code{groupingVar} parameters are desired, especially of rate, it is highly recommended to use
\code{\link{TEfit}} or \code{\link{TEfitAll}}.

The \code{formIn} must include a random effect of \code{timeVar} by \code{groupingVar}
(e.g., \code{(time_variable | grouping_variable)})
}
\examples{
\dontrun{
m_TElmem <- TElmem(acc ~ trialNum + (trialNum || subID), anstrain, timeVar = 'trialNum',groupingVar = 'subID')
# Typical lmer model:
summary(m_TElmem$lmerMod) # On average, starting accuracy was .137 worse than asymptotic accuracy
# Participant-level rate parameters:
m_TElmem$rates
}
}
