% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anstrain_s1}
\alias{anstrain_s1}
\title{Perceptual learning data from one participant}
\format{An object of class \code{data.frame} with 250 rows and 6 columns.}
\usage{
anstrain_s1
}
\description{
Data from a participant completing 250 trials each of perceptual learning.
"resp" is participant response to the "ratio" offset, with correctness recorded
as "acc". "sizeRat" is a covariate.
}
\keyword{datasets}
