% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TEfit.R
\name{plot.TEfit}
\alias{plot.TEfit}
\title{Plot a TEfit}
\usage{
\method{plot}{TEfit}(TEs3, plot_title = "", xlabel = "", ylabel = "",
  sub_title = "", ymin = NA, ymax = NA)
}
\arguments{
\item{TEs3}{{{EXPLAIN}}}

\item{plot_title}{{{EXPLAIN}}}

\item{xlabel}{{{EXPLAIN}}}

\item{ylabel}{{{EXPLAIN}}}

\item{sub_title}{{{EXPLAIN}}}

\item{ymin}{{E}}

\item{ymax}{{E}}
}
\description{
Plot a TEfit
}
