% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.TEfit.R
\name{simulate.TEfit}
\alias{simulate.TEfit}
\title{Simulate from a TEfit, assuming multivariate gaussian parameters}
\usage{
\method{simulate}{TEfit}(object, nsim = 100, seed = NULL, newdata = data.frame(), toPlot = F)
}
\arguments{
\item{object}{A TEfit model}

\item{nsim}{number of simulations to generate}

\item{seed}{Null, to retain symmetry with other simulate() methods}

\item{newdata}{If desired, new data from which to generate values}

\item{toPlot}{Logical, for plotting of values.}
}
\description{
Requires the $psych$ and $MASS$ packages.
}
\examples{
\dontrun{simulate(model_fit_by_TEfit)}

}
