% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEbrm.R
\name{TEbrm}
\alias{TEbrm}
\title{Fit a time-evolving model with Stan using brms}
\usage{
TEbrm(
  formIn,
  dataIn,
  ...,
  iter = 1000,
  chains = 3,
  priorIn = c(),
  algorithm = "sampling",
  link_start_asym = "",
  tef_control_list = TEfits::tef_control()
)
}
\arguments{
\item{formIn}{Formula to fit. See examples.}

\item{dataIn}{Data frame, from which to fit the model.}

\item{...}{Further arguments passed to the brms model}

\item{iter}{Number of iterations to run the model.}

\item{chains}{Number of chains to run the model.}

\item{priorIn}{Optional argument to pass priors to the \code{brms} model, alongside the TEfit-default rate prior. If you provide any, you will likely need to provide priors for all nonlinear parameters. \code{brm} error messages tend to be very helpful in this regard.}

\item{algorithm}{The algorithm to use when fitting the \code{\link[brms]{brm}} model}

\item{link_start_asym}{Link function to use for the start and asymptote parameters. Defaults to what is passed from formIn. Otherwise, most likely to be 'log' or 'inv_logit'}

\item{tef_control_list}{A list of control parameters passed in by \code{tef_control()}}
}
\description{
Formats and runs a \code{\link[brms]{brm}s} model for a time-evolving nonlinear function. Function is \strong{under development}
and is likely to be buggy, and to change frequently.
}
\details{
When specifying statistical families, it is \emph{extremely highly recommended} to use an "identity" link function,
and then [if appropriate] specifying a link function using \code{link_start_asym}.

\code{formIn} is a formula, with the time-varying response variable on the left, followed by \code{~}.
The right side must be either [A] a single variable corresponding to the dimension of time, or
[B] a call to a \code{TEfits} constructor function such as \code{\link{tef_change_expo3}}. See
examples.

Currently supported model constructor functions  are:
\itemize{
\item{\code{tef_change_expo3} -- 3-parameter exponential (start, [inverse] rate, and asymptote) -- rate is log of time to some proportion remaining, default is log2 of time to 50 percent remaining}
}
}
\note{
Default priors and parameter boundaries are implemented, but all users would benefit from
re-fitting models with different priors in order to ensure that inferences are not biased
by defaults. The assumptions that guided the creation of the default priors may not be
appropriate for your data. If no \code{link_start_asym} function is used (i.e., the default
'identity') then start and asymptote priors are Gaussian with the response variable's mean, and double
the response variable's SD. If a \code{link_start_asym} function is used (e.g., 'log' or
'inv_logit') then start and asymptote priors are Gaussian with a mean of zero and a SD of 3 (which
may place quite a bit of the prior density at values more extreme than appropriate for many
users' data). Default [log time constant] rate parameter's prior is Gaussian centered at the
log of the mean of the time variable (with the base of the log defined in \code{tef_control_list}).
The SD of this prior is 1/3 of the mean, and boundaries are implemented at extreme values. All
other priors are \code{\link[brms]{brm}} defaults.
Use \code{brms::prior_summary} to extract priors from a fitted model object.
}
\examples{
\dontrun{
## Default model formula is exponential change, with no covariates or random effects
m <- TEbrm(
acc ~ trialNum # equivalent to `acc ~ tef_change_expo3('trialNum')`
,priorIn = prior(normal(.5,.5),nlpar='pAsym') + prior(normal(.5,.5),nlpar='pStart')
,dataIn = anstrain_s1
)

prior_summary(m)
summary(m)
conditional_effects(m)

## using the tef_change_expo3 function to construct the model formula, with random effects
m <- TEbrm(
acc ~ tef_change_expo3('trialNum',parForm = ~ (1|subID))
,dataIn = anstrain
,priorIn = prior(normal(.5,.5),nlpar='pAsym') + prior(normal(.5,.5),nlpar='pStart')
)

## Estimate accuracy using a more appropriate [bernoulli] response function,
## ## and also estimate the start and asymptote parameters using invert-logit links
m <- TEbrm(
acc ~ tef_change_expo3('trialNum',parForm = ~ (1|subID))
,dataIn = anstrain
,link_start_asym = 'inv_logit'
,family=bernoulli(link='identity')
)

}
}
\seealso{
For additional flexibility, and full explanations of model options, see \code{\link[brms]{brms-package}}.
}
