% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEbrm.R
\name{TEbrm}
\alias{TEbrm}
\title{Fit a time-evolving model with Stan using brms}
\usage{
TEbrm(
  formIn,
  dataIn,
  ...,
  iter = 1000,
  chains = 3,
  priorIn = c(),
  algorithm = "sampling",
  link_start_asym = "",
  tef_control_list = TEfits::tef_control()
)
}
\arguments{
\item{formIn}{A formula, with the time-varying response variable on the left, followed by \code{~}.  The right side must be either [A] a single variable corresponding to the dimension of time, or [B] a call to a \code{TEfits} constructor function such as \code{\link{tef_change_expo3}}. See examples.}

\item{dataIn}{Data frame, from which to fit the model.}

\item{...}{Further arguments passed to the brms model}

\item{iter}{Number of iterations to run the model.}

\item{chains}{Number of chains to run the model.}

\item{priorIn}{Optional argument to pass priors to the \code{brms} model, alongside the TEfit-default rate prior. If you provide any, you will likely need to provide priors for all nonlinear parameters. \code{brm} error messages tend to be very helpful in this regard.}

\item{algorithm}{The algorithm to use when fitting the \code{\link[brms]{brm}} model}

\item{link_start_asym}{Inverse of the link function to use for the start and asymptote parameters. Defaults to what is passed from formIn. Otherwise, the user would most likely to want to use 'exp' or 'inv_logit'.}

\item{tef_control_list}{A list of control parameters passed in by \code{tef_control()}}
}
\description{
Formats and runs a \code{\link[brms]{brm}s} model for a time-evolving nonlinear function. Function is \strong{under development}
and is likely to be buggy, and to change frequently.
}
\details{
When specifying statistical families, it is \emph{extremely highly recommended} to specify an "identity" link function,
and then [if appropriate] specifying a link function using the \code{link_start_asym} argument. See example.

Currently supported model constructor functions are:
\itemize{
\item{\code{tef_change_expo3} -- 3-parameter exponential (start, [inverse] rate, and asymptote) -- rate is log of time to some proportion remaining, default is log2 of time to 50 percent remaining}
}

Currently supported link functions are:
\itemize{
\item{\code{tef_link_logistic} -- logistic psychometric function, parameterized in terms of threshold and asymptote. Takes the output of a \code{tef_change_} function and augments it prior to use by \code{TEbrm}}
}
}
\note{
Default priors and parameter boundaries are implemented, but all users would benefit from
re-fitting models with different priors in order to ensure that inferences are not biased
by defaults. The assumptions that guided the creation of the default priors may not be
appropriate for your data. If no \code{link_start_asym} function is used (i.e., the default
'identity') then start and asymptote priors are Gaussian with the response variable's mean, and double
the response variable's SD. If a \code{link_start_asym} function is used (e.g., 'log' or
'inv_logit') then start and asymptote priors are Gaussian with a mean of zero and a SD of 3 (which
may place quite a bit of the prior density at values more extreme than appropriate for many
users' data). Default [log time constant] rate parameter's prior is Gaussian centered at the
log of the mean of the time variable (with the base of the log defined in \code{tef_control_list}).
The SD of this prior is 1/3 of the mean, and boundaries are implemented at extreme values. All
other priors are \code{\link[brms]{brm}} defaults.
Use \code{\link[brms]{prior_summary}} to examine priors from a fitted model object; see
\code{\link[brms]{set_prior}} for setting priors.
}
\examples{
\dontrun{

## Default model formula is exponential change, with no covariates or random effects
m1 <- TEbrm(
acc ~ trialNum # equivalent to `acc ~ tef_change_expo3('trialNum')`
,dataIn = anstrain_s1
)

prior_summary(m1)
summary(m1)
conditional_effects(m1)

## using the tef_change_expo3 function to construct the model formula, with random effects
m2 <- TEbrm(
acc ~ tef_change_expo3('trialNum',parForm = ~ (1|subID))
,dataIn = anstrain
,priorIn = prior(normal(.5,.5),nlpar='pAsym') + prior(normal(.5,.5),nlpar='pStart')
)

## Estimate accuracy using a more appropriate [bernoulli] response function,
## # and also estimate the start and asymptote parameters using invert-logit links
m3 <- TEbrm(
acc ~ tef_change_expo3('trialNum',parForm = ~ (1|subID))
,dataIn = anstrain
,link_start_asym = 'inv_logit'
,family=bernoulli(link='identity')
)

## Fit a time-evolving logistic mixed-effects model (see, e.g., Cochrane et al., 2019, AP&P, 10.3758/s13414-018-01636-w).
## # May take a few minutes to run.
m4 <- TEbrm(
resp ~ tef_link_logistic(
tef_change_expo3('trialNum', parForm = ~ (1|subID))
, linkX = 'ratio' )
,family=bernoulli(link='identity')
,iter = 4000
,dataIn = anstrain
)

}
}
\seealso{
For additional flexibility, and full explanations of model options, see \code{\link[brms]{brms-package}}.

For other approaches to time-evolving models, see \code{\link{TEfits}}.
}
