% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEbrm_advi.R
\name{TEbrm_advi}
\alias{TEbrm_advi}
\title{Run a brm model with ADVI}
\usage{
TEbrm_advi(
  formIn,
  dataIn = data.frame(),
  ...,
  algorithm = "fullrank",
  conv_thresh = 0.1,
  quiet = F
)
}
\arguments{
\item{formIn}{Model formula, as in \code{\link[brms]{brm}}.}

\item{dataIn}{Data, as in \code{\link[brms]{brm}}.}

\item{...}{Any other argument to pass to \code{\link[brms]{brm}}.}

\item{algorithm}{which ADVI algorithm to use: "meanfield" or "fullrank".}

\item{quiet}{Progress is printed by default, but can be suppressed with quiet=T.}
}
\description{
Uses Stan's stochastic gradient ascent methods "fullrank" or "meanfield" rather than full
Bayesian sampling. Is likely to be much faster than typical sampling, but somewhat
less accurate. This function fits the model several times and returns the output
with the highest Bayesian R-squared. \strong{Beware: This
has a tendency to crash R sometimes.}
}
\details{
Stochastic gradient ascent in Stan uses Automatic Differentiation Variational Inference (ADVI).
}
\note{
Re-fitting a model within this function is not comprehensive. If using ADVI, it is recommended
to use \code{TEbrm_advi} multiple times, and choose the best using a comparison such as
\code{\link[brms]{add_criterion}(model,'loo')}.
}
\examples{
\dontrun{
## can be used in place of brm
m <- TEbrm_advi(ratio ~ resp, anstrain_s1)
summary(m)
conditional_effects(m)

## Use in the context of TEfits
m1 <- TEbrm(
acc ~ tef_change_expo3('trialNum')
,dataIn = anstrain_s1,
,algorithm = 'meanfield'
)

}
}
\seealso{
\code{\link[rstan]{vb}} and \link{http://mc-stan.org}
}
