% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_change_expo3.R
\name{tef_change_expo3}
\alias{tef_change_expo3}
\title{Construct a 3-parameter exponential function of change}
\usage{
tef_change_expo3(
  timeVar,
  parForm = ~1,
  startForm = ~1,
  rateForm = ~1,
  asymForm = ~1,
  changeBase = 2,
  rateBase = 2
)
}
\arguments{
\item{timeVar}{String. The name of the variable in the model that corresponds to time. The variable of time should be positive and numeric, and the function of change should be expected to happen with increasing time.}

\item{parForm}{The right-hand side of the formula defining all nonlinear parameters as well as the null [non-time-varying] model.}

\item{startForm}{The right-hand side of the formula defining the start parameter. Overwrites \code{parForm} for this parameter.}

\item{rateForm}{The right-hand side of the formula defining the rate parameter. Overwrites \code{parForm} for this parameter.}

\item{asymForm}{The right-hand side of the formula defining the asymptote parameter. Overwrites \code{parForm} for this parameter.}

\item{changeBase}{Number. The base of the log (e.g., 2 or exp(1)) of the change function.}

\item{rateBase}{Number. The base of the log (e.g., 2 or exp(1)) of the rate [time constant].}
}
\description{
By defining the model variable associated with time (e.g., trial number), and
formulas defining each of the nonlinear parameters of time-related change,
this function constructs a model that can then be passed to functions for fitting
the model (e.g., \code{\link{TEbrm}}).
}
\details{
Function is \strong{under development}
and is likely to be buggy, and to change frequently.
}
\examples{
equation_to_fit <- tef_change_expo3('timeVar',rateForm = ~ xvar1*xvar2) # both variables should be numeric for TEfit methods! TEbrm should work with factorsas well

equation_to_fit <- tef_change_expo3('timeVar'
, parForm = ~ xvar1   # overall parameter formula; overwritten for time-evolving formulas below, leaving this as just the null model
, startForm = ~ xvar2 # start parameter's regression model
, rateForm = ~ (1|participantID)  # rate parameter's [mixed-effects] regression model
, asymForm = ~ xvar3 # asymptote parameter's regression model
)

}
\seealso{
\code{\link{TEbrm}} for examples of how to use this function in specifying models.
}
