% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_checkPars.R
\name{tef_checkPars}
\alias{tef_checkPars}
\title{Check for bound hitting and other undesireable outcomes within an optim() call}
\usage{
tef_checkPars(err, guesses, curDat, pNames, evalFun, errFun, respVar,
  linkFunX = NA, y_lim, rate_lim, shape_lim, penalizeRate, paramTerms,
  guessGroups = guessGroups)
}
\arguments{
\item{err}{Error}

\item{guesses}{Parameter values}

\item{curDat}{Data being fit}

\item{pNames}{Parameter names}

\item{evalFun}{Function being fit}

\item{errFun}{Function to calculate error}

\item{respVar}{Name of the response variable}

\item{linkFunX}{If relevant, the "x" value for a link function (e.g., Weibull, logistic)}

\item{y_lim}{Limits to fit values}

\item{rate_lim}{Limits to rate parameter}

\item{shape_lim}{If using a Weibull change function, limits to Weibull shape parameter}

\item{penalizeRate}{Logical. Should error be penalized if rate is extremely close to the bounds?}

\item{paramTerms}{parameter-level regressions, to be evaluated for checking y_lim and rate_lim}

\item{guessGroups}{deprecated}
}
\description{
TEfits internal
}
