% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_control.R
\name{tef_control}
\alias{tef_control}
\title{Get control parameters for a TEfit model}
\usage{
tef_control(
  quietErrs = F,
  suppressWarnings = F,
  nTries = 200,
  y_lim = c(-1e+07, 1e+07),
  rate_lim = c(0, 0),
  shape_lim = c(0, 0),
  expBase = 2,
  rateBase = 2,
  pFix = c(),
  penalizeMean = T,
  penalizeRate = F,
  convergeTol = 0.05,
  stepwise_asym = F,
  explicit = ""
)
}
\arguments{
\item{quietErrs}{logical. Should errors be printed to the Console?}

\item{suppressWarnings}{logical. Should warnings be printed to the Console?}

\item{nTries}{Numeric. What is the maximum number of optimization runs that should be attempted?}

\item{y_lim}{Numeric vector of length 2. Lower and upper bounds of permitted predicted values.}

\item{rate_lim}{Numeric vector of length 2. Lower and upper bounds of permitted rate values [log time constants].}

\item{shape_lim}{Numeric vector of length 2. Lower and upper bounds of permitted shape parameter values (i.e., for Weibull).}

\item{expBase}{For change functions with an exponential component, what should the base of the exponent be?}

\item{rateBase}{What should the base of the rate exponent be?}

\item{pFix}{Named numeric vector allowing specific parameters to be fixed to a constant (i.e., not estimated)}

\item{penalizeMean}{Logical. Should the time-evolving model be penalized if the mean of the time-evolving predicted values diverges from the mean of the null [non-time-evolving] predicted values?}

\item{penalizeRate}{Logical. Should the time-evolving model be penalized if the rate parameter is very near a boundary?}

\item{convergeTol}{Convergence is extremely roughly defined in \code{TEfits} as the SD of the same estimated parameter on different runs with relatively low error. What should this SD be?}

\item{stepwise_asym}{Logical. If a function will saturate by the end of the measurement time, this option allows the asymptote to be estimated from this time period (i.e., as stationary).}

\item{explicit}{Character. Rather than using any of the pre-defined change or link functions, enter the specific function you want to test.}
}
\description{
\code{\link{TEfit}} internal
}
