% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_fitAll2brms.R
\name{tef_fitAll2brms}
\alias{tef_fitAll2brms}
\title{Refit a TEfitAll model with brms}
\usage{
tef_fitAll2brms(
  TEs3s,
  nIter = 2000,
  nChains = 3,
  nCores = 2,
  errFun = NA,
  prior_dispersion = 2
)
}
\arguments{
\item{TEs3s}{TEfitAll model}

\item{nIter}{number of iterations}

\item{nChains}{number of chains}

\item{nCores}{number of cores}

\item{errFun}{the error function to use. Defaults to the same as the TEfitAll model, if possible.}

\item{prior_dispersion}{This number, multiplied by the SD of each TEfitAll parameter, is used as the prior SD for that parameter.}
}
\value{
A \code{\link[brms]{brms}} nonlinear mixed-effects model object.
}
\description{
Passes a \code{\link{TEfitAll}} model to [nonlinear mixed-effects Bayesian] fitting using the
\code{\link[brms]{brms}} package. Note that, due to the extensive time needed to
fit \code{\link[brms]{brms}} models,
this function is less tested than most functions in the \code{TEfits} package. Functionality is
\strong{experimental}.
}
\details{
\code{TEfitAll} \code{bernoulli} models are fit using either \code{bernoulli} or \code{Beta} response
distributions in \code{brms} depending on whether the \code{TEfitAll} distribution is
binary. \code{TEfitAll} \code{logcosh} models are fit using a \code{asym_laplace} response distribution
in brms.
}
\note{
Under development. Partial functionality.
}
