% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_fitAll2brms.R
\name{tef_fitAll2brms}
\alias{tef_fitAll2brms}
\title{Refit a TEfitAll model with brms}
\usage{
tef_fitAll2brms(TEs3s, nIter = 2000, nChains = 3, nCores = 2)
}
\arguments{
\item{TEs3s}{TEfitAll model}

\item{nIter}{number of iterations}
}
\description{
Refit a TEfitAll model with brms
}
