% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_fitAll2brms.R
\name{tef_fitAll2brms}
\alias{tef_fitAll2brms}
\title{Refit a TEfitAll model with brms}
\usage{
tef_fitAll2brms(
  TEs3s,
  fixef = NA,
  nIter = 2000,
  nChains = 3,
  nCores = 2,
  errFun = NA,
  prior_dispersion = 2
)
}
\arguments{
\item{TEs3s}{TEfitAll model}

\item{fixef}{Parameters vary as random effects by the TEs3s grouping variable. However, if you have main effects (e.g., group differences), enter them \emph{as a data frame} here.}

\item{nIter}{number of iterations}

\item{nChains}{number of chains}

\item{nCores}{number of cores}

\item{errFun}{the error function to use. Defaults to the same as the TEfitAll model, if possible.}

\item{prior_dispersion}{This number, multiplied by the SD of each TEfitAll parameter, is used as the prior SD for that parameter.}
}
\value{
A \code{\link[brms]{brms-package}} nonlinear mixed-effects model object.
}
\description{
Passes a \code{\link{TEfitAll}} model to [nonlinear mixed-effects Bayesian] fitting using
\code{\link[brms]{brms-package}}. Note that, due to the extensive time needed to
fit \code{\link[brms]{brms-package}} models,
this function is less tested than most functions in the \code{TEfits} package. Functionality is
\strong{experimental}.
}
\details{
Priors for nonlinear parameters are informed by the distributions of parameters in the \code{TEfitAll} object [models].
However, any fixed effects should be minimally influenced by these priors

\code{TEfitAll} \code{bernoulli} models are fit using either \code{bernoulli} or \code{Beta} response
distributions in \code{brms} depending on whether the \code{TEfitAll} distrIibution is
binary. \code{TEfitAll} \code{logcosh} models are fit using a \code{asym_laplace} response distribution
in brms predicting the .5 quantile.
}
\note{
Under development. Partial functionality.
}
\examples{
\dontrun{
dat <- anstrain
dat$condition <- rep(c('A','B'),each=500)

# Model with time and one categorical fixed effect
mod_tef <- TEfitAll(dat[,c('acc','trialNum')], groupingVar = dat$subID)
mod_brm <- tef_fitAll2brms(mod_tef,nChains=1,fixef=data.frame(condition=dat$condition))

# Model with time, one categorical fixed effect, and one by-groupingVar (subID) random slope
dat$absRat <- scale(abs(dat$ratio))
mod_tef <- TEfitAll(dat[,c('acc','trialNum',"absRat")], groupingVar = dat$subID,covarTerms=list(pRate=c(F)))
mod_brm <- tef_fitAll2brms(mod_tef,nChains=1,fixef=data.frame(condition=dat$condition))
}

}
