% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_fitErr.R
\name{tef_fitErr}
\alias{tef_fitErr}
\title{Calculate the error of data, given the model and certain parameters}
\usage{
tef_fitErr(
  guesses,
  varIn,
  pNames,
  evalFun,
  errFun,
  respVar,
  linkFunX = NA,
  y_lim,
  rate_lim,
  shape_lim,
  penalizeMean,
  penalizeRate,
  parLims,
  thresh_fun,
  paramTerms
)
}
\arguments{
\item{guesses}{proposed parameter values}

\item{varIn}{Data to be fit}

\item{pNames}{Parameter names}

\item{evalFun}{Model function}

\item{errFun}{Error function}

\item{respVar}{The name of the response variable}

\item{linkFunX}{variable used in link function}

\item{y_lim}{Bounds for the predicted values}

\item{rate_lim}{Bounds for the rate parameter}

\item{shape_lim}{Bounds for the shape parameter, if needed.}

\item{penalizeMean}{two-element vector. The first element is logical, the second is the mean predicted value from the null fit}

\item{penalizeRate}{logical, whether or not to penalize rate parameter}

\item{parLims}{parameter limits}

\item{thresh_fun}{if relevant, the threshold function}

\item{paramTerms}{the terms for parameters}
}
\description{
\code{\link{TEfit}} internal.
}
