% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getBounds.R
\name{tef_getBounds}
\alias{tef_getBounds}
\title{Define parameter boundaries and guessing ranges}
\usage{
tef_getBounds(modList, whichPnames = "pNames", linkFunX = NA)
}
\arguments{
\item{modList}{TEfit modList}

\item{whichPnames}{which parameter set?}

\item{linkFunX}{If applicable, the variable along which the link function varies}
}
\description{
\code{\link{TEfit}} internal. Future versions will systematize this more thoroughly.
Designed to variation in the starting points for optimization runs.
}
