% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getDatTerms.R
\name{tef_getDatTerms}
\alias{tef_getDatTerms}
\title{For a TEfit model, use a change function to gt the appropriate RHS terms}
\usage{
tef_getDatTerms(modList, pPrefix, whichChange = "expo")
}
\arguments{
\item{modList}{List of TEfit model details}

\item{pPrefix}{{{CREATE EXPLANATION}}}

\item{whichChange}{{{CREATE EXPLANATION}}}
}
\description{
For a TEfit model, use a change function to gt the appropriate RHS terms
}
