% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsExpo.R
\name{tef_getTermsExpo}
\alias{tef_getTermsExpo}
\title{Define a time-dependent change function}
\usage{
tef_getTermsExpo(modList, pPrefix)
}
\arguments{
\item{modList}{TEfit modList}

\item{pPrefix}{parameter prefix}
}
\description{
\code{\link{TEfit}} internal. Note that, despite the name, this is the getterms function for  *both* exponential and power [3-parameter] change functions
}
