% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsExpoDouble.R
\name{tef_getTermsExpoDouble}
\alias{tef_getTermsExpoDouble}
\title{Construct parameter terms for the double-rate change function}
\usage{
tef_getTermsExpoDouble(modList, pPrefix)
}
\arguments{
\item{modList}{List of TEfit model details}

\item{pPrefix}{prefix for parameters (e.g., p or thresh)}
}
\description{
NOTE: test more thoroughly.
}
