% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsExpoFatigue.R
\name{tef_getTermsExpoFatigue}
\alias{tef_getTermsExpoFatigue}
\title{Define a time-dependent change function}
\usage{
tef_getTermsExpoFatigue(modList, pPrefix)
}
\arguments{
\item{modList}{TEfit modList}

\item{pPrefix}{parameter prefix}
}
\description{
Internal to TEfits. Exponential change followed by logistic decay.
}
