% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tef_getTermsPower4.R
\name{tef_getTermsPower4}
\alias{tef_getTermsPower4}
\title{Define a time-dependent change function}
\usage{
tef_getTermsPower4(modList, pPrefix)
}
\arguments{
\item{modList}{TEfit modList}

\item{pPrefix}{parameter prefix}
}
\description{
\code{\link{TEfit}} internal. 4-parameter power function.
}
