% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEfitAll.R
\name{TEfitAll}
\alias{TEfitAll}
\title{Fit several time-evolving regression models}
\usage{
TEfitAll(varIn, groupingVar, returnAll = T, progressDot = T,
  linkFun = list(link = "identity"), errFun = "ols",
  changeFun = "expo", bootPars = list(nBoots = 0, bootTries = 0,
  bootPercent = 0), blockTimeVar = NULL, covarTerms = list(),
  control = list())
}
\arguments{
\item{varIn}{Data frame or vector. First column [or vector] is the time-dependent response variable. If available, second column is the time variable. All other columns are covariates, possibly involved in a link function.}

\item{groupingVar}{Variable (e.g., participant ID) with which to separate TEfit models. Length must be nrows(varIn)}

\item{returnAll}{Logical. Return only a summary (when T), or that summary plus every model, in a list (when F)}

\item{progressDot}{If TRUE, prints a dot after each group fit}

\item{linkFun}{A list defining a link function (i.e., 'identity', 'Quick', or 'logistic')}

\item{errFun}{A string defining an error function (e.g., 'ols', 'logcosh', 'exgauss'). Full description {here}}

\item{changeFun}{A string defining the functional form of change (e.g., 'expo', 'power', 'weibull')}

\item{bootPars}{A list defining the details for bootstrapped fits. Defaults to no bootstrapping. Necessary for estimates of uncertainty around fits and for covariance between parameters.}

\item{blockTimeVar}{A string defining the time points of sub-scales (e.g., "blocks" of times within the overall timescale of data collection)}

\item{covarTerms}{{{description}}}

\item{control}{A list of model parameters. Full description {here}}
}
\description{
A wrapper for fitting a TEfit model to the data
for every unique value of groupingVar. Defaults to
returning a list including two summaries and all models; returning
only a summary is also an option.
}
