% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEglmem.R
\name{TEglmem}
\alias{TEglmem}
\title{Generalized linear mixed-effects model with nonlinear time random effects}
\usage{
TEglmem(
  formIn,
  dat,
  timeVar,
  groupingVar,
  family = gaussian,
  startingOffset = T,
  nRuns = 1,
  silent = F
)
}
\arguments{
\item{formIn}{model formula, as in \code{glmer}}

\item{dat}{model data, as in \code{glmer}}

\item{timeVar}{String. Indicates which variable in \code{datIn} corresponds to time (i.e., should be transformed). Must be numeric and positive.}

\item{groupingVar}{String. Indicates which variable in \code{datIn} should have a time-related random effect.}

\item{family}{model family, as in \code{glmer}}

\item{startingOffset}{By default (if T) time is coded to start at 1 and saturate to 0. If startingOffset is F, time starts at 0 and saturates to 1. May assist in interpreting interactions with other variables, etc.}

\item{nRuns}{Number of times to run optimization of the rate (i.e., fitting nonlinear transformations of \code{timeVar})}

\item{silent}{Progress is printed by default. silent=T to suppress}
}
\value{
A list including:
\describe{
\item{\code{glmerMod}}{\code{\link[lme4]{glmer}} model fit with transformed time variable}
\item{\code{rates}}{Named vector of rates [\emph{binary log of 50-percent-of-change time constants}]}
\item{\code{timeDat}}{Data frame with original and transformed time variable}
\item{\code{groupMods}}{List of fit \code{\link{TEglm}} models, and the corresponding transformed time variable and named vector of rates}
}
}
\description{
Fits a \code{\link[lme4]{glmer}} generalized linear mixed-effects model with the random effects of
\code{timeVar} for each level of \code{groupingVar}. Provides estimates of time-related change
(i.e., attempts to answer the question "how different was the start than the end?").
}
\details{
First uses \code{\link{TEglm}} to find a rate parameter for each level of \code{groupingVar}, with
the formula extracted using \code{\link{tef_getRanefForm}}. These
rate parameters are used to transform the corresponding \code{timeVar} into a exponentially
saturating variable (see \code{\link{TEglm}}). After finding an initial set of
rate parameters using \code{\link{TEglm}},
\code{TEglmem} attempts to optimize the vector of rate parameters in conjunction with the full
\code{glmer} model.

May be used, with \code{nRuns=0}, as a wrapper for \code{\link{TEglm}} in order to simply
use rate estimates from independent \code{groupingVar}-level models, extract the corresponding
transformed time variable, and use this in a GLMEM.
}
\note{
Random effects and rate estimates may be unstable, and optimization may take
a very long time to run. The primary purpose of this function is to allow for by-\code{groupingVar}
detrending of time-related changes in data (i.e., to estimate and test fixed effects at asymptotic time,
or to estimate and test the magnitude of time-related effects).
If reliable by-\code{groupingVar} parameters are desired, it is highly recommended to use
\code{\link{TEfit}} or \code{\link{TEfitAll}}.

The \code{formIn} must include a random effect of \code{timeVar} by \code{groupingVar}
(e.g., \code{(time_variable | grouping_variable)}).

Although the time variable is transformed to exponentially decay toward zero, this does not necessarily mean
that the model prediction involves an exponential change with time. The nonlinear change in time relates
to the time-associated model coefficients.
}
\examples{
\dontrun{
m_TEglmem <- TEglmem(resp ~ ratio + trialNum:ratio + (ratio + trialNum:ratio || subID),anstrain, timeVar = 'trialNum',groupingVar = 'subID',family=binomial)
# Typical glmer model:
summary(m_TEglmem$glmerMod)
# Participant-level rate parameters:
m_TEglmem$rates
}

}
