% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEhelp.R
\name{TEhelp}
\alias{TEhelp}
\title{Get help for the TEfits package}
\usage{
TEhelp(topic = "package")
}
\arguments{
\item{topic}{Help topic.}
}
\description{
Many options for fitting using TEfits are specified in obscure lists and such. To help,
this function provides verbose explanations of how to use various tools.
}
\examples{
TEhelp('package')
TEhelp('errFun')
TEhelp('linkFun=d_prime')
TEhelp('errFun=logcosh')

}
