% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEhelp.R
\name{TEhelp}
\alias{TEhelp}
\title{Get help for the TEfits package}
\usage{
TEhelp(topic = "package")
}
\arguments{
\item{topic}{Help topic. To see options, see TEhelp('package')}
}
\description{
Get help for the TEfits package
}
