% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TElm.R
\name{TElm}
\alias{TElm}
\title{[Robust] linear model with nonlinear time predictor}
\usage{
TElm(formIn, dat, timeVar, robust = F, fixRate = NA, nBoot = 250)
}
\arguments{
\item{formIn}{model formula, as in \code{lm()}}

\item{dat}{model data, as in \code{lm()}}

\item{timeVar}{String. Indicates which model predictor is time (i.e., should be transformed)}

\item{robust}{Logical. Should \code{\link[MASS]{rlm}} be used?}

\item{fixRate}{If numeric, use this as a rate parameter [binary-log of 50 percent time constant] rather than estimating it (e.g., to improve reproducibility)}

\item{nBoot}{Number of bootstrapped models to fit after rate [time constant] has been estimated (passed to \code{\link{tef_rlm_boot}})}
}
\description{
Fit a linear model or robust linear model with time as a covariate,
while estimating the shape of the nonlinear interpolation between starting and ending time.
First resamples data with replacement 200 times, and each time estimates the best-shaped curve to interpolate
between initial time-related offset and asymptotic time (i.e., rate at which effect of time saturates at zero).
Then uses the mean estimated rate to transform the \code{timeVar} predictor into an exponentially decaying variable
interpolating between initial time (time offset magnitude of 1) and arbitrarily large time values (time
offset magnitude 0). Last uses this transformed time variable in a \code{\link{tef_rlm_boot}} [\code{rlm} or \code{lm}] model
(i.e., attempts to answer the question "how different was the start than the end?").
}
\details{
Rate is parameterized as a time constant, or the amount of time it takes for half of change to occur.
The value of rate has a lower bound of
the .0333 quantile of the time variable (i.e., 87.5\% of change happens in the first 10\% of time) and an upper bound of the
.333 quantile of the time variable (i.e., 87.5\% of change takes 100\% of the time to happen). These bounds provide
some robustness in estimates of asympototic effects (i.e., "controlling for time") as well as initial effects
(i.e., "time-related starting offset"). Uses this transformed time variable in a \code{\link{tef_rlm_boot}} model to estimate
bootstrapped parameter coefficients and out-of-sample prediction.

Mean estimated rate is calculated after trimming the upper 25% and lower 25% of bootstrapped rate estimates, for robustness to
extremes in resampling.
}
\examples{
dat <- data.frame(trialNum = 1:200, resp = log(11:210)+rnorm(200))

m_lm <- TElm(resp ~ trialNum,dat, 'trialNum')
summary(m_lm)
m_lm$bootSummary
m_lm$rate

m_rlm <- TElm(resp ~ trialNum,dat,'trialNum',robust=TRUE)
summary(m_rlm)
m_rlm$bootSummary
m_rlm$rate

plot(dat[,c('trialNum','resp')])
lines(dat$trialNum,fitted(m_lm),col='blue')
lines(dat$trialNum,fitted(m_rlm),col='red')

}
\seealso{
\code{\link{TElmem}} for mixed-effects extension of \code{TElm};
\code{\link{TEglm}} for generalized extension of \code{TElm}
}
