% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TElmem.R
\name{TElmem}
\alias{TElmem}
\title{Linear mixed-effects model with nonlinear time random effects}
\usage{
TElmem(formIn, dat, timeVar, groupingVar, onlyGroupMods = F, nRuns = 5)
}
\arguments{
\item{formIn}{model formula, as in \code{lmer}}

\item{dat}{model data, as in \code{lmer}}

\item{timeVar}{String. Indicates which variable in \code{datIn} corresponds to time (i.e., should be transformed)}

\item{groupingVar}{String. Indicates which variable in \code{datIn} should have a time=related random effect.}

\item{onlyGroupMods}{IF TRUE, returns only the by-\code{groupingVar} fits from \code{\link{TElm}}}

\item{nRuns}{Number of times to run optimization of the rate (i.e., fitting nonlinear transformations of \code{timeVar})}
}
\description{
Fits a \code{lme4::lmer} linear mixed-effects model with the random effects of
\code{timeVar} for each level of \code{groupingVar}.
}
\details{
First uses \code{\link{TElm}} to find a rate parameter for each level of \code{groupingVar}. These
rate parameters are used to transform the corresponding \code{timeVar} into a exponentially
saturating variable (see \code{\link{TElm}}). After finding bivariate rate parameters using \code{\link{TElm}},
\code{TElmem} attempts to optimize the vector of rate parameters in conjunction with the full
\code{lmer} model. Random effects and rate estimates may be unstable, and optimization may take
a very long time to run. The primary purpose of this function is to allow for by-\code{groupingVar}
detrending of time-related changes in data (i.e., to estimate and test fixed effects at asymptotic time).
If reliable by-\code{groupingVar} parameters are desired, it is highly recommended to use
\code{TEfit} or \code{TEfitAll}.

May be used, with \code{onlyGroupMods=T}, as a wrapper for \code{\link{TElm}} in order to simply
fit bivariate \code{response~time} models and extract the corresponding transformed time variable.

Note that the \code{formIn} must include a random effect of \code{timeVar} by \code{groupingVar}
(e.g., \code{(time_variable | grouping_variable)}).
}
